# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RPay Payment Gateway API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices (admin sees all, merchant sees own)
      parameters:
        - name: merchantId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice for a merchant (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/mark-paid:
    post:
      operationId: markInvoicePaid
      tags: [invoices]
      summary: Mark invoice as paid (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice marked paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/void:
    post:
      operationId: voidInvoice
      tags: [invoices]
      summary: Void an invoice (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice voided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
components:
  schemas:
...
    Invoice:
      type: object
      required: [id, merchantId, invoiceNumber, amount, currency, status, createdAt]
      properties:
        id:
          type: integer
        merchantId:
          type: integer
        merchantName:
          type: ["string", "null"]
        merchantEmail:
          type: ["string", "null"]
        planId:
          type: ["integer", "null"]
        planName:
          type: ["string", "null"]
        invoiceNumber:
          type: string
        amount:
          type: string
        currency:
          type: string
        period:
          type: ["string", "null"]
        periodFrom:
          type: ["string", "null"]
        periodTo:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, issued, paid, void]
        dueDate:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    InvoiceInput:
      type: object
      required: [merchantId, amount]
      properties:
        merchantId:
          type: integer
        planId:
          type: ["integer", "null"]
        amount:
          type: string
        currency:
          type: string
        period:
          type: ["string", "null"]
        periodFrom:
          type: ["string", "null"]
        periodTo:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
    InvoiceListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...