# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPay Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceStatus } from './invoiceStatus';

export interface Invoice {
  id: number;
  merchantId: number;
  /** @nullable */
  merchantName?: string | null;
  /** @nullable */
  merchantEmail?: string | null;
  /** @nullable */
  planId?: number | null;
  /** @nullable */
  planName?: string | null;
  invoiceNumber: string;
  amount: string;
  currency: string;
  /** @nullable */
  period?: string | null;
  /** @nullable */
  periodFrom?: string | null;
  /** @nullable */
  periodTo?: string | null;
  status: InvoiceStatus;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}