# lib/api-zod/src/generated/types/invoiceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPay Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceInput {
  merchantId: number;
  /** @nullable */
  planId?: number | null;
  amount: string;
  currency?: string;
  /** @nullable */
  period?: string | null;
  /** @nullable */
  periodFrom?: string | null;
  /** @nullable */
  periodTo?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  notes?: string | null;
  status?: string;
}