# lib/api-zod/src/generated/types/invoiceListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPay Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { Invoice } from './invoice';

export interface InvoiceListResponse {
  data: Invoice[];
  total: number;
  page: number;
  limit: number;
}